Move ResultAssertions to test package
