+++ ground/src/test/java/com/google/android/ground/ResultAssertions.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.sharedtest
+package com.google.android.ground
 
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
